requirements.txt
Flask==2.3.2
Pillow==10.0.1
numpy==1.26.4
pyyaml==6.0
python-multipart==0.0.5